+++ notarook-ie/polykeys.h
@@ -1,0 +1,17 @@
+/**
+ * This file is for handling the keys section of the polyglot
+ * book openings format
+ */
+
+#ifndef POLYKEYS_H
+#define POLYKEYS_H
+
+#ifdef _MSC_VER
+#define U64(u) (u##ui64)
+#else
+#define U64(u) (u##ULL)
+#endif
+
+extern const uint64_t Random64[781];
+
+#endif